requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beanie==1.26.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.6.1
exceptiongroup==1.2.0
fastapi==0.105.0
h11==0.14.0
idna==3.6
Jinja2==3.1.4
lazy-model==0.2.0
MarkupSafe==2.1.5
motor==3.5.1
pydantic==2.5.2
pydantic_core==2.14.5
pymongo==4.8.0
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
toml==0.10.2
typing_extensions==4.9.0
urllib3==2.1.0
uvicorn==0.24.0.post1
